Add Basic Categories [With Logic]
This Commit adds more Categories on the left of the Screen below the Login Button (TAS Playback, TAS Old/Beta). All of these have a hidden Pane (the one from the right) in another Pane, that can switch between them whenever clicking them. Those Panes contain all Gamemodes and/or Informations like Patch notes and Warnings. We still have to add the logic behind all the buttons on these Panes and create new Subpanes for 'Installations' and 'Patch notes' which themselves contain Information that has to be Downloaded from a Server that doesn't exist yet. This Server may be a normal HTTP Server, or a Custom written Server. I don't know yet.