Rename 'TAS Battle Launcher' to 'TAS Launcher'
This is now a Launcher that can Launch all kinds of TAS Stuff. My Plan was making this a TAS Battle Launcher so that people can easily play TAS Battle. But now that I think about it I can just make it run every single Minecraft Version and act as a Launcher with automatic Mod Install.