Automatically download the latest Mod Data from the Server
I added all the missing Webpages and now they are all connected to the Launcher. Beneath the Description there now will be the latest Updates from the Mods (they need to be reworked on the Server though). The Launcher will also display a Message if it is offline for some reason. This will later do some more stuff